- 메인 프레그 먼트 , 스토리 , 송, 책장 및 단어장 추가 - 뉴스 프레그먼트 추가 - 관련 뷰모델 추가

+++ app/src/main/java/com/littlefox/app/foxschool/common/Common.kt
@@ -75,8 +75,8 @@ class Common
         const val SERVICE_NOT_ENDED : String    = "N"
         const val SERVICE_ENDED : String        = "Y"
 
-        const val TARGET_PHONE_DISPLAY_WIDTH : Int = 1080
-        const val TARGET_TABLET_DISPLAY_WIDTH : Int = 1920
+        const val TARGET_PHONE_DISPLAY_WIDTH : Float = 1080.0f
+        const val TARGET_TABLET_DISPLAY_WIDTH : Float = 1920.0f
 
         const val LOADING_DIALOG_SIZE : Int              = 150
